pyproject.toml
[project]
name = "safari-writer"
version = "0.1.3"
description = "A UI-compatible clone of AtariWriter, built with Textual"
requires-python = ">=3.10"
authors = [
    { name = "Matthew Martin", email = "matthewdeanmartin@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
keywords = ["editor", "retro"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "textual>=1.0.0",
    "pyenchant>=3.2.0",
    "pygments>=2.17.0",
    "mastodon-py>=2.1.4",
    "python-dotenv>=1.0",
    "send2trash>=2.1.0",
    "httpx>=0.27.0",
    "dbf>=0.99.11",
    "gitpython>=3.1.0",
    "Pillow>=10.0.0",
    "term-image>=0.7.0",
    "reportlab>=4.4.10",
]

[project.scripts]
safari-writer = "safari_writer.main:main"
safari-dos = "safari_dos.main:main"
safari-chat = "safari_chat.main:main"
safari-base = "safari_base.main:main"
safari-fed = "safari_fed.main:main"
safari-asm = "safari_asm.main:main"
safari-repl = "safari_repl.main:main"
safari-reader = "safari_reader.main:main"
safari-view = "safari_view.main:main"
safari-view-tk = "safari_view.ui_tk.tk_app:main"

[project.urls]
"Bug Tracker" = "https://github.com/matthewdeanmartin/safari-writer/issues"
"Change Log" = "https://github.com/matthewdeanmartin/safari-writer/blob/main/CHANGELOG.md"
"Documentation" = "https://safari-writer.readthedocs.io/en/latest/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["safari_writer", "safari_dos", "safari_chat", "safari_base", "safari_fed", "safari_basic", "safari_asm", "safari_repl", "safari_reader", "safari_view"]
include = [
    "safari_writer/locales/LC_MESSAGES/*.mo",
    "safari_writer/locales/LC_MESSAGES/*.po",
    "safari_writer/*.sfw",
    "safari_writer/*.json",
    "safari_writer/locales/**/*.mo",
    "safari_writer/macros/*.bas",
    "safari_writer/py.typed",
    "safari_dos/py.typed",
    "safari_chat/default_help.md",
    "safari_chat/py.typed",
    "safari_base/py.typed",
    "safari_fed/py.typed",
    "safari_asm/example/*.asm",
    "safari_asm/py.typed",
    "safari_repl/py.typed",
    "safari_reader/py.typed",
    "safari_view/py.typed",
    "/README.md", "LICENSE",
]





[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pylint>=3.3.0",
    "ruff>=0.11.0",
    "textual-dev>=1.0.0",
    "pytest>=8.0.0",
    "hypothesis>=6.151.9",
    "tox>=4.0.0",
    "hatch",
    "types-pygments>=2.19.0.20251121",
    "pytest-cov>=7.0.0",
]

[tool.mypy]
python_version = "3.12"
files = ["safari_writer"]
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = ["enchant"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["reportlab", "reportlab.*"]
ignore_missing_imports = true
